requirements.txt
# Core dependencies for AI Course Assignment
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
seaborn>=0.11.0
pillow>=8.0.0
scipy>=1.7.0

# Reinforcement Learning
gymnasium>=1.1.0
pygame>=2.6.1

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: For better development experience
tqdm>=4.62.0
